+++ src/Factories/Factory.php
@@ -18,6 +18,7 @@
 
 namespace CloudCreativity\JsonApi\Factories;
 
+use CloudCreativity\JsonApi\Contracts\Encoder\SerializerInterface;
 use CloudCreativity\JsonApi\Contracts\Factories\FactoryInterface;
 use CloudCreativity\JsonApi\Contracts\Http\ApiInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Requests\RequestInterpreterInterface;
@@ -27,8 +28,11 @@ use CloudCreativity\JsonApi\Contracts\Store\StoreInterface;
 use CloudCreativity\JsonApi\Contracts\Validators\QueryValidatorInterface;
 use CloudCreativity\JsonApi\Encoder\Encoder;
 use CloudCreativity\JsonApi\Http\Api;
+use CloudCreativity\JsonApi\Http\Client\GuzzleClient;
 use CloudCreativity\JsonApi\Http\Query\ValidationQueryChecker;
 use CloudCreativity\JsonApi\Http\Requests\RequestFactory;
+use CloudCreativity\JsonApi\Http\Responses\Response;
+use CloudCreativity\JsonApi\Object\Document;
 use CloudCreativity\JsonApi\Pagination\Page;
 use CloudCreativity\JsonApi\Repositories\CodecMatcherRepository;
 use CloudCreativity\JsonApi\Repositories\ErrorRepository;
@@ -43,7 +47,11 @@ use Neomerx\JsonApi\Contracts\Http\Headers\SupportedExtensionsInterface;
 use Neomerx\JsonApi\Contracts\Schema\ContainerInterface as SchemaContainerInterface;
 use Neomerx\JsonApi\Encoder\EncoderOptions;
 use Neomerx\JsonApi\Factories\Factory as BaseFactory;
+use Psr\Http\Message\MessageInterface;
+use Psr\Http\Message\ResponseInterface as PsrResponse;
 use Psr\Http\Message\ServerRequestInterface;
+use function CloudCreativity\JsonApi\http_contains_body;
+use function CloudCreativity\JsonApi\json_decode;
 
 /**
  * Class Factory
@@ -54,11 +62,17 @@ class Factory extends BaseFactory implements FactoryInterface
 {
 
     /**
-     * @param SchemaContainerInterface $container
-     * @param EncoderOptions|null $encoderOptions
-     * @return Encoder
+     * @inheritdoc
      */
     public function createEncoder(SchemaContainerInterface $container, EncoderOptions $encoderOptions = null)
+    {
+        return $this->createSerializer($container, $encoderOptions);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function createSerializer(SchemaContainerInterface $container, EncoderOptions $encoderOptions = null)
     {
         $encoder = new Encoder($this, $container, $encoderOptions);
         $encoder->setLogger($this->logger);
@@ -100,6 +114,35 @@ class Factory extends BaseFactory implements FactoryInterface
         $requestFactory = new RequestFactory($this);
 
         return $requestFactory->build($httpRequest, $intepreter, $api);
+    }
+
+
+    /**
+     * @inheritDoc
+     */
+    public function createResponse(PsrResponse $response)
+    {
+        return new Response($response, $this->createDocumentObject($response));
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function createDocumentObject(MessageInterface $message)
+    {
+        if (!http_contains_body($message)) {
+            return null;
+        }
+
+        return new Document(json_decode($message->getBody()));
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function createClient($httpClient, SchemaContainerInterface $container, SerializerInterface $encoder)
+    {
+        return new GuzzleClient($this, $httpClient, $container, $encoder);
     }
 
     /**
@@ -185,7 +228,6 @@ class Factory extends BaseFactory implements FactoryInterface
 
         return new ValidationQueryChecker($checker, $validator);
     }
-
     /**
      * @inheritDoc
      */